Added support for controlling generation of shared bin libs or not and profile updates made for minimum mac os host machine version support (10.10 and 10.12)
(1) Added shared option in create-pkg.py script to specify if libs should
be shared or not.

(2) Code updates made to various modules to manage generating dynamic or
static bin libs.

(3) Updates made to Mac OS profiles to reduce minimum required version numbers
of supported Mac OS host machines to 10.10 and 10.12.

(4) Added support for specifying if builds should generate shared bin libs
or not is only currently active in the recipes libzamq and protobuf.